fix(build): 스토리북에 query provider 추가 (#89)
