--- .storybook/preview.ts
@@ -1,16 +1,0 @@
-import type {Preview} from '@storybook/nextjs';
-import '@/app/globals.css';
-
-const preview: Preview = {
-  parameters: {
-    controls: {
-      matchers: {
-        color: /(background|color)$/i,
-        date: /Date$/i,
-      },
-    },
-    layout: 'centered',
-  },
-};
-
-export default preview;
